requirements.txt
absl-py==2.0.0
accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyascii==0.3.2
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
av==10.0.0
azure-cognitiveservices-speech==1.32.1
Babel==2.13.1
bangla==0.0.2
blinker==1.7.0
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
cachetools==5.3.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
clean-fid==0.1.35
click==8.1.7
clip-anytorch==2.5.2
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.2.0
contourpy==1.2.0
coqpit==0.0.17
ctranslate2==3.20.0
cycler==0.12.1
Cython==0.29.30
dateparser==1.1.8
dctorch==0.1.2
decorator==5.1.1
diskcache==5.6.3
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.1
elevenlabs==0.2.26
encodec==0.1.1
enum34==1.1.10
executing==2.0.1
faster-whisper==0.9.0
filelock==3.9.0
Flask==2.3.3
flatbuffers==23.5.26
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
g2pkk==0.1.2
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.2
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
halo==0.0.31
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
imageio==2.31.6
inflect==5.6.2
ipython==8.17.2
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonlines==1.2.0
jsonmerge==1.9.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
k-diffusion==0.1.1
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.3
librosa==0.10.1
llama_cpp_python==0.2.74
llvmlite==0.40.1
log-symbols==0.0.14
Markdown==3.5.1
MarkupSafe==2.1.2
matplotlib==3.7.3
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.57.0
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.16.1
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathtools==0.1.2
Pillow==9.3.0
platformdirs==3.11.0
pooch==1.8.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.6
pure-eval==0.2.2
pvporcupine==1.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pynndescent==0.5.10
pyparsing==3.1.1
pypinyin==0.49.0
pypiwin32==223
pyreadline3==3.4.1
pysbd==0.3.4
python-crfsuite==0.9.9
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
RealTimeSTT==0.1.7
RealTimeTTS==0.2.2
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.12.0
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.3
sentry-sdk==1.34.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.3.7
spinners==0.0.24
stack-data==0.6.3
stream2sentence==0.1.6
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.13.3
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchdiffeq==0.2.3
torchsde==0.2.6
torchvision==0.16.0+cu118
tqdm==4.64.1
trainer==0.0.31
traitlets==5.13.0
trampoline==0.1.2
transformers==4.33.3
TTS==0.19.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
umap-learn==0.5.4
Unidecode==1.3.7
urllib3==1.26.13
wandb==0.15.12
wcwidth==0.2.9
webrtcvad==2.0.10
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.2
py-cpuinfo==9.0.0
